+++ src/main/java/phillip/stockinfo4j/model/pojo/StockDailyTran.java
@@ -25,10 +25,10 @@ public class StockDailyTran implements Serializable {
     private String name;
 
     @Column(name = "trading_vol",nullable = false)
-    private Integer tradingVol = 0;//成交張數
+    private Long tradingVol = 0L;//成交張數
 
     @Column(name = "deal",nullable = false)
-    private Integer deal = 0;//成交筆數
+    private Long deal = 0L;//成交筆數
 
     @Column(name = "opening",nullable = false)
     private Double opening = 0.00 ;//開盤價
